display players bets on the table change float to double in most places, 222.18 was becoming 22217 after *100

+++ qt-client/client/table/table_sit.h
@@ -28,8 +28,8 @@ private slots:
 	void On_seat_taken(QByteArray gameid);
 	void On_PlayerClubStatus(Data::PlayerClubStatus &pcs);
 private:
-	float GetBuyinMin();
-	float GetBuyinMax();
+	double GetBuyinMin();
+	double GetBuyinMax();
 	void updateLimits();
 
 	Ui::TableSit *ui;
